fix: comprehensive audit remediation — security hardening, code quality, and modularization
Security:
- Remove all hardcoded rustbrain_dev_2024 credentials; fail fast on missing env vars
- Restrict query_graph endpoint with parameterized queries and read-only Neo4j user
- Cap depth parameter to max 10 with validation
- Bind database ports to 127.0.0.1 only in docker-compose
- Apply escapeHtml() to all API-sourced data in playground HTML
- Sanitize error messages returned to clients (no internal details leaked)
- Redact database URLs from startup logs

Code Quality:
- Split 2098-line api/main.rs into modules (config, errors, handlers/, neo4j, state)
- Fix gaps.rs functional bug (as_array on JSON objects → as_object)
- Replace std::thread::sleep with tokio::time::sleep in async contexts
- Remove blanket #![allow(dead_code)] and address individual warnings
- Commit Cargo.lock for reproducible builds (remove *.lock from .gitignore)
