+++ .env.example
@@ -1,8 +1,8 @@
 # =============================================================================
 # rust-brain — Environment Configuration
 # =============================================================================
-# Copy to .env and customize for your environment.
-# All values below are development defaults — change passwords for production.
+# Copy to .env and fill in real values before running.
+# Placeholder values (<...>) MUST be replaced — services will fail to start otherwise.
 # =============================================================================
 
 # --- Project ---
@@ -10,20 +10,25 @@ COMPOSE_PROJECT_NAME=rustbrain
 
 # --- Postgres ---
 POSTGRES_USER=rustbrain
-POSTGRES_PASSWORD=rustbrain_dev_2024
+POSTGRES_PASSWORD=<your-password>
 POSTGRES_DB=rustbrain
 POSTGRES_PORT=5432
-POSTGRES_URL=postgresql://rustbrain:rustbrain_dev_2024@postgres:5432/rustbrain
+POSTGRES_URL=postgresql://rustbrain:<your-password>@postgres:5432/rustbrain
 
 # --- Pgweb ---
 PGWEB_PORT=8081
 
 # --- Neo4j ---
-NEO4J_AUTH=neo4j/rustbrain_dev_2024
-NEO4J_PASSWORD=rustbrain_dev_2024
+NEO4J_AUTH=neo4j/<your-password>
+NEO4J_PASSWORD=<your-password>
 NEO4J_HTTP_PORT=7474
 NEO4J_BOLT_PORT=7687
 NEO4J_URI=bolt://neo4j:7687
+# Read-only user for the API service (query_graph endpoint).
+# Password must match the CREATE USER statement in scripts/init-neo4j.cypher.
+# On Neo4j Enterprise, grant the reader role to enforce read-only at the DB level.
+NEO4J_READONLY_USER=rustbrain_readonly
+NEO4J_READONLY_PASSWORD=<your-readonly-password>
 
 # --- Qdrant ---
 QDRANT_REST_PORT=6333
