+++ .gitignore
@@ -6,5 +6,4 @@
 .idea/
 .vscode/
 target/
-*.lock
 target-repo/
