+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
     image: postgres:16-alpine
     container_name: rustbrain-postgres
     ports:
-      - "${POSTGRES_PORT:-5432}:5432"
+      - "127.0.0.1:${POSTGRES_PORT:-5432}:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
       - ./scripts/init-db.sql:/docker-entrypoint-initdb.d/01-init.sql
@@ -38,7 +38,7 @@ services:
     image: sosedoff/pgweb:latest
     container_name: rustbrain-pgweb
     ports:
-      - "${PGWEB_PORT:-8081}:8081"
+      - "127.0.0.1:${PGWEB_PORT:-8081}:8081"
     environment:
       PGWEB_DATABASE_URL: "postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres:5432/${POSTGRES_DB}?sslmode=disable"
     depends_on:
@@ -51,8 +51,8 @@ services:
     image: neo4j:5-community
     container_name: rustbrain-neo4j
     ports:
-      - "${NEO4J_HTTP_PORT:-7474}:7474"
-      - "${NEO4J_BOLT_PORT:-7687}:7687"
+      - "127.0.0.1:${NEO4J_HTTP_PORT:-7474}:7474"
+      - "127.0.0.1:${NEO4J_BOLT_PORT:-7687}:7687"
     volumes:
       - neo4j_data:/data
       - neo4j_logs:/logs
@@ -78,8 +78,8 @@ services:
     image: qdrant/qdrant:v1.12.1
     container_name: rustbrain-qdrant
     ports:
-      - "${QDRANT_REST_PORT:-6333}:6333"
-      - "${QDRANT_GRPC_PORT:-6334}:6334"
+      - "127.0.0.1:${QDRANT_REST_PORT:-6333}:6333"
+      - "127.0.0.1:${QDRANT_GRPC_PORT:-6334}:6334"
     volumes:
       - qdrant_data:/qdrant/storage
     environment:
@@ -213,8 +213,8 @@ services:
     environment:
       DATABASE_URL: postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres:5432/${POSTGRES_DB}
       NEO4J_URI: bolt://neo4j:7687
-      NEO4J_USER: neo4j
-      NEO4J_PASSWORD: ${NEO4J_PASSWORD}
+      NEO4J_USER: ${NEO4J_READONLY_USER:-rustbrain_readonly}
+      NEO4J_PASSWORD: ${NEO4J_READONLY_PASSWORD}
       QDRANT_HOST: http://qdrant:6333
       QDRANT_COLLECTION: code_embeddings
       OLLAMA_HOST: http://ollama:11434
