+++ services/api/src/handlers/playground.rs
@@ -1,0 +1,14 @@
+//! Playground route handler.
+
+use axum::{
+    http::StatusCode,
+    response::IntoResponse,
+};
+
+pub async fn playground_html() -> impl IntoResponse {
+    (
+        StatusCode::OK,
+        [("Content-Type", "text/html; charset=utf-8")],
+        include_str!("../../static/playground.html"),
+    )
+}
