+++ services/ingestion/src/parsers/syn_parser.rs
@@ -95,6 +95,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -127,6 +128,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -159,6 +161,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -219,6 +222,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -270,6 +274,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -302,6 +307,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -332,6 +338,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -362,6 +369,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -392,6 +400,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -421,6 +430,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
     
@@ -454,6 +464,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
 
@@ -488,6 +499,7 @@ impl SynParser {
             start_line: skeleton.start_line,
             end_line: skeleton.end_line,
             body_source: source.to_string(),
+            generated_by: None,
         })
     }
