+++ tests/integration/test_pipeline.sh
@@ -38,11 +38,11 @@ fi
 # Default values
 POSTGRES_PORT="${POSTGRES_PORT:-5432}"
 POSTGRES_USER="${POSTGRES_USER:-rustbrain}"
-POSTGRES_PASSWORD="${POSTGRES_PASSWORD:-rustbrain_dev_2024}"
+POSTGRES_PASSWORD="${POSTGRES_PASSWORD:-<your-password>}"
 POSTGRES_DB="${POSTGRES_DB:-rustbrain}"
 NEO4J_HTTP_PORT="${NEO4J_HTTP_PORT:-7474}"
 NEO4J_BOLT_PORT="${NEO4J_BOLT_PORT:-7687}"
-NEO4J_PASSWORD="${NEO4J_PASSWORD:-rustbrain_dev_2024}"
+NEO4J_PASSWORD="${NEO4J_PASSWORD:-<your-password>}"
 QDRANT_REST_PORT="${QDRANT_REST_PORT:-6333}"
 
 # Helper function to print test result
